Resolve code review suggestions #21
