+++ src/core/utility.cpp
@@ -16,24 +16,14 @@ std::string core::StrToUpper(std::string str) {
     return str;
 }
 
-std::string core::StrTrimWhitespace(std::string str) {
-    auto is_whitespace_char = [](const char c) {
-        switch (c) {
-        case ' ':
-        case '\0':
-            return true;
-
-        default:
-            return false;
-        }
-    };
-
+std::string core::StrTrim(std::string str, const std::vector<char>& chars) {
+    auto is_trim_char = [&](const char c) { return std::find(chars.begin(), chars.end(), c) != chars.end(); };
 
     // Trim leading
     {
         std::size_t str_start = 0; // Where actual string begins (non whitespace)
 
-        while (is_whitespace_char(str[str_start]) && str_start < str.size()) {
+        while (is_trim_char(str[str_start]) && str_start < str.size()) {
             ++str_start;
         }
 
@@ -44,7 +34,7 @@ std::string core::StrTrimWhitespace(std::string str) {
     if (!str.empty()) {
         std::size_t str_end = str.size() - 1; // Actual string ends
 
-        while (is_whitespace_char(str[str_end]) && str_end > 0) {
+        while (is_trim_char(str[str_end]) && str_end > 0) {
             --str_end;
         }
 
@@ -52,6 +42,10 @@ std::string core::StrTrimWhitespace(std::string str) {
     }
 
     return str;
+}
+
+std::string core::StrTrimWhitespace(const std::string& str) {
+    return StrTrim(str, {' ', '\0'});
 }
 
 std::string core::StrMatchLen(std::string str, const std::size_t target_len, const char padding) {
