--- src/render/gui/gui_renderer.cpp
@@ -1,5 +1,0 @@
-// This file is subject to the terms and conditions defined in 'LICENSE' included in the source code package
-
-#include "render/gui/gui_renderer.h"
-
-using namespace jactorio;
